% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix-process.R
\name{filter_low}
\alias{filter_low}
\title{Title}
\usage{
filter_low(mat, thres_count = 2, thres_sample_nums = 5)
}
\arguments{
\item{mat}{integer matrix.}

\item{thres_sample_nums}{}
}
\description{
Title
}
