% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization.R
\name{norm_cpm}
\alias{norm_cpm}
\alias{norm_cpm_top}
\alias{norm_cpm_rm}
\alias{norm_cpm_refer}
\title{CPM normalization}
\usage{
norm_cpm(mat)

norm_cpm_top(mat, top_n)

norm_cpm_rm(mat, gene_type)

norm_cpm_refer(mat, refer_gene_id)
}
\arguments{
\item{mat}{integer matrix. counts.}

\item{gene_type}{character.}

\item{refer_gene_id}{character.}
}
\description{
CPM normalization
}
\details{
some functions may throw errors
}
\section{Functions}{
\itemize{
\item \code{norm_cpm}: normalize by all genes

\item \code{norm_cpm_top}: normalize by top k genes sorted by expression level

\item \code{norm_cpm_rm}: normalize by removing genes of given type (such as piRNAs)

\item \code{norm_cpm_refer}: normalize by of given reference genes
}}

\examples{
norm_cpm_refer(sim_mat, suggest_refer$id)

}
