% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization.R
\name{norm_cpm_impl}
\alias{norm_cpm_impl}
\title{CPM normalization by some genes}
\usage{
norm_cpm_impl(mat, row)
}
\arguments{
\item{mat}{integer matrix. counts}

\item{row}{integer or logical. Use which rows (genes) as normalization factor}
}
\description{
CPM normalization by some genes
}
