% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization.R
\name{norm_scater}
\alias{norm_scater}
\alias{norm_tmm}
\alias{norm_rle}
\title{TMM/RLE normalization}
\usage{
norm_tmm(mat)

norm_rle(mat)
}
\arguments{
\item{mat}{integer matrix. counts}
}
\description{
TMM/RLE normalization
}
\section{Functions}{
\itemize{
\item \code{norm_tmm}: TMM normalization

\item \code{norm_rle}: RLE normalization
}}

