% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization.R
\name{norm_SCnorm}
\alias{norm_SCnorm}
\title{SCnorm normalization}
\usage{
norm_SCnorm(mat, ...)
}
\arguments{
\item{mat}{integer matrix. counts.}

\item{...}{other arguments passed on to \code{\link[SCnorm:SCnorm]{SCnorm::SCnorm()}}.}
}
\value{
numerical matrix. normalized counts.
}
\description{
SCnorm normalization
}
\examples{
norm_SCnorm(sim_mat*10)

}
\seealso{
Other matrix normalization: \code{\link{norm_cpm}},
  \code{\link{norm_scater}}
}
\concept{matrix normalization}
