% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization.R
\name{normalize}
\alias{normalize}
\title{martix normalization}
\usage{
normalize(counts_mat_path, norm_methods = c("SCnorm", "TMM", "RLE",
  "CPM", "CPM_top", "CPM_rm", "CPM_refer"), top_n = NULL,
  rm_gene_type = NULL, refer_gene_id = NULL,
  sample_class_path = NULL, PCA_label_by = NULL, PAC_color_by = NULL,
  refer_gene_name = refer_gene_id, output_dir = ".",
  output_file = "norm")
}
\description{
martix normalization
}
\examples{
\\donotrun{
    refer_gene_id = c(
        "ENST00000408438.1", "ENST00000385271.1", "ENST00000607334.3",
        "ENST00000385059.1", "ENST00000362134.1", "ENST00000385245.1",
        "ENST00000385045.1", "ENST00000362117.1", "ENST00000384832.1",
        "ENST00000579846.3"
    )
    norm_mat(
        '/path/to/matrix', c('SCnorm', 'TMM', 'RLE', 'CPM', 'CPM_top', 'CPM_rm', 'CPM_refer'),
        top_n = 20L, rm_gene_type = c('miRNA', 'piRNA'), refer_gene_id = refer_gene_id
    )
}
}
