% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix-process.R
\name{plot_cv_density}
\alias{plot_cv_density}
\title{density plot of coefficient of variation}
\usage{
plot_cv_density(mat, refer_gene_id = "",
  refer_gene_name = refer_gene_id)
}
\arguments{
\item{mat}{integer matrix. counts}

\item{refer_gene_id}{character. Ensembl transcript id, add a vertical line for each gene to mark the corresponding CV (on x axis). Only genes in counts matrix would be shown. Usually these genes should be the reference genes you want to use for normalization.}

\item{refer_gene_name}{character. Transcript name}
}
\value{
\code{\link[ggplot2:ggplot]{ggplot2::ggplot()}} object
}
\description{
density plot of coefficient of variation
}
\examples{


}
