% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix-process.R
\name{plot_group_impl}
\alias{plot_group_impl}
\title{workhorse of \code{plot_PCA/TSNE()}}
\usage{
plot_group_impl(sce, shape = NULL, color = NULL, plot_fun)
}
\description{
workhorse of \code{plot_PCA/TSNE()}
}
\details{
\code{plot_PCA(sce, shape, color) -> }plot_group_impl(sce, shape, color, scater::plotPCA)`
}
\keyword{internal}
