% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix-process.R
\name{plot_PCA}
\alias{plot_PCA}
\alias{plot_TSNE}
\title{plot PCA, TSNE}
\usage{
plot_PCA(sce, shape = NULL, color = NULL)

plot_TSNE(sce, shape = NULL, color = NULL)
}
\arguments{
\item{sce}{A SingleCellExperiment object.}

\item{shape, }{color string. specify a column in \code{col_data} of \code{\link[=as_SingleCellExperiment]{as_SingleCellExperiment()}} to shape/color by}
}
\description{
plot PCA, TSNE
}
\section{Functions}{
\itemize{
\item \code{plot_TSNE}: 
}}

