% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix-process.R
\name{plot_variance}
\alias{plot_variance}
\alias{plot_cv_density}
\alias{plot_refer_violin}
\title{plot variance of counts across samples}
\usage{
plot_cv_density(mat, refer_gene_id = "",
  refer_gene_name = refer_gene_id)

plot_refer_violin(mat, refer_gene_id, refer_gene_name = refer_gene_id)
}
\arguments{
\item{mat}{integer matrix. counts}

\item{refer_gene_id}{character. Ensembl transcript id, add a vertical line
for each gene to mark the corresponding CV (on x axis). Only genes in
counts matrix would be shown. Usually these genes should be the reference
genes you want to use for normalization.}

\item{refer_gene_name}{character. Transcript name}
}
\value{
\code{\link[ggplot2:ggplot]{ggplot2::ggplot()}} object.
}
\description{
plot variance of counts across samples
}
\details{
\code{plot_cv_density()} produces density plot of coefficient of variation

\code{plot_cv_density()} produces density plot of coefficient of variation
}
\examples{
# only one gene exist in the matrix
plot_cv_density(sim_mat, suggest_refer$id)
plot_cv_density(sim_mat, suggest_refer$id, suggest_refer$name)

# the name should be the same length as id
plot_cv_density(sim_mat, rownames(sim_mat)[1:6], letters[1:3])
# if only part of the genes have name, you can pass the id of other genes
plot_cv_density(sim_mat, rownames(sim_mat)[1:6], c(letters[1:3], rownames(sim_mat)[4:6]))

# only one gene exist in the matrix
plot_refer_violin(sim_mat, suggest_refer$id)
plot_refer_violin(sim_mat, suggest_refer$id, suggest_refer$name)

# the name should be the same length as id
plot_refer_violin(sim_mat, rownames(sim_mat)[1:6], letters[1:3])
# if only part of the genes have name, you can pass the id of other genes
plot_refer_violin(sim_mat, rownames(sim_mat)[1:6], c(letters[1:3], rownames(sim_mat)[4:6]))

}
