% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix-process.R
\name{read_mat}
\alias{read_mat}
\title{read count matrix}
\usage{
read_mat(path, ...)
}
\arguments{
\item{path}{string.}

\item{...}{other arguments passsed on to \code{\link[readr:read_tsv]{readr::read_tsv()}}}
}
\value{
integer matrix
}
\description{
read count matrix
}
\examples{
NULL
}
