% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix-process.R
\name{seq_y}
\alias{seq_y}
\title{generate equally spaced y coordinates, not hit bottom nor top.}
\usage{
seq_y(plot, x)
}
\arguments{
\item{plot}{ggplot object.}

\item{x}{numberic vector}
}
\value{
numeric vector. The same length as \code{x}
}
\description{
generate equally spaced y coordinates, not hit bottom nor top.
}
\details{
Image \code{plot}'s y axis extends from 0 to 3, \code{x} contains 3 values,
then we give \code{c(0.5, 1.5, 2.5)}.
}
\keyword{internal}
