% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix-process.R
\name{filter_low}
\alias{filter_low}
\title{filter genes with low expression values}
\usage{
filter_low(mat, min_count = 2, min_sample_per_gene = 5)
}
\arguments{
\item{mat}{integer matrix.}

\item{min_count, }{min_sample_per_gene integer scalar. For each gene, it must
contain at least \code{min_count} reads in at least \code{min_sample_per_gene}
samples. Otherwise, it would be dropped.}
}
\value{
integer matrix.
}
\description{
filter genes with low expression values
}
\examples{
filter_low(sim_mat)

}
