% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization.R
\name{norm_cpm}
\alias{norm_cpm}
\title{CPM normalization}
\usage{
norm_cpm(mat)
}
\arguments{
\item{mat}{integer matrix. counts}
}
\description{
CPM normalization
}
\section{Functions}{
\itemize{
\item \code{norm_cpm}: normalize by all genes
}}

