% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization.R
\name{norm_cpm}
\alias{norm_cpm}
\alias{norm_cpm_total}
\alias{norm_cpm_top}
\alias{norm_cpm_rm}
\alias{norm_cpm_refer}
\title{CPM normalization}
\usage{
norm_cpm_total(mat)

norm_cpm_top(mat, top_n)

norm_cpm_rm(mat, gene_type)

norm_cpm_refer(mat, refer_gene_id)
}
\arguments{
\item{mat}{integer matrix. counts.}

\item{top_n}{integer scalar. see \code{norm_cpm_top()} below}

\item{gene_type}{character. see \code{norm_cpm_rm()} below}

\item{refer_gene_id}{character. Ensembl transcript id, see \code{norm_cpm_refer()} below}
}
\description{
CPM normalization using counts sum of \emph{certain} genes as scaling factor
}
\details{
some functions may throw errors

\code{norm_cpm_total()} uses total genes

\code{norm_cpm_top()} uses top 20 genes sorted by counts (assuming \code{top_n = 20L})

\code{norm_cpm_rm()} uses non-piRNA genes (assuming \code{gene_type = 'piRNA'})

\code{norm_cpm_refer()} uses given reference genes
}
\examples{
norm_cpm_total(sim_mat)

norm_cpm_top(sim_mat, 20L)

norm_cpm_rm(sim_mat, c('miRNA', 'piRNA'))

norm_cpm_refer(sim_mat, suggest_refer$id)

}
\seealso{
Other matrix normalization: \code{\link{norm_SCnorm}},
  \code{\link{norm_scater}}
}
\concept{matrix normalization}
