% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization.R
\name{norm_scater}
\alias{norm_scater}
\alias{norm_tmm}
\alias{norm_rle}
\title{TMM/RLE normalization by scater package}
\usage{
norm_tmm(mat)

norm_rle(mat)
}
\arguments{
\item{mat}{integer matrix. counts.}
}
\value{
numerical matrix. normalized counts.
}
\description{
TMM/RLE normalization by scater package
}
\details{
\code{norm_tmm()} performs TMM normalization.

\code{norm_rle()} performs RLE normalization.
}
\examples{
norm_tmm(sim_mat)

norm_rle(sim_mat)

}
\seealso{
Other matrix normalization: \code{\link{norm_SCnorm}},
  \code{\link{norm_cpm}}
}
\concept{matrix normalization}
