% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix-process.R
\name{plot_cv_density}
\alias{plot_cv_density}
\title{Title}
\usage{
plot_cv_density(mat, refer_gene_id = "",
  refer_gene_name = refer_gene_id)
}
\arguments{
\item{refer_gene_name}{}
}
\description{
Title
}
