% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix-process.R
\name{read_mat}
\alias{read_mat}
\title{Read counts matrix}
\usage{
read_mat(path, ...)
}
\arguments{
\item{path}{string.}

\item{...}{other arguments passsed on to \code{\link[readr:read_tsv]{readr::read_tsv()}}}
}
\value{
integer matrix
}
\description{
Read counts matrix
}
\details{
In any case, first part (separated by \code{|}) of row names must be
Ensembl transcript id
}
